% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createDummyData}
\alias{createDummyData}
\title{Create Dummy Data}
\usage{
createDummyData()
}
\description{
Returns a randomly generated experience table for 100 policies,
for 2016-2019.  By default the rows will be ordered byy policy start date
}
\examples{
ed1 <- createDummyData()

}
