% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{earlierDate}
\alias{earlierDate}
\title{Earlier Date}
\usage{
earlierDate(dt1, dt2)
}
\arguments{
\item{dt1}{a vector of dates}

\item{dt2}{a vector of dates}
}
\description{
Returns the earlier of two dates
}
\details{
Applies element-wise to vectors of dates
}
\examples{

}
\seealso{
Other Date Functions: \code{\link{isBetweenDates}},
  \code{\link{laterDate}}
}
\concept{Date Functions}
