% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isBetweenDates}
\alias{isBetweenDates}
\title{Is Between Dates}
\usage{
isBetweenDates(dtx, dt1, dt2)
}
\arguments{
\item{dtx}{a vector of dates}

\item{dt1}{a vector of dates}

\item{dt2}{a vector of dates}
}
\description{
Returns true if date X is between dates 1 and 2.
}
\details{
Applies element-wise to vectors of dates
}
\examples{

}
\seealso{
Other Date Functions: \code{\link{earlierDate}},
  \code{\link{laterDate}}
}
\concept{Date Functions}
